+++ broker/src/main/java/io/camunda/zeebe/broker/partitioning/startup/steps/SnapshotStoreStep.java
@@ -8,7 +8,7 @@
 package io.camunda.zeebe.broker.partitioning.startup.steps;
 
 import io.camunda.zeebe.broker.partitioning.startup.PartitionStartupContext;
-import io.camunda.zeebe.db.impl.rocksdb.ChecksumProvider;
+import io.camunda.zeebe.db.impl.rocksdb.ChecksumProviderRocksDBImpl;
 import io.camunda.zeebe.scheduler.SchedulingHints;
 import io.camunda.zeebe.scheduler.future.ActorFuture;
 import io.camunda.zeebe.scheduler.startup.StartupStep;
@@ -29,7 +29,7 @@ public final class SnapshotStoreStep implements StartupStep<PartitionStartupCont
         new FileBasedSnapshotStore(
             context.partitionMetadata().id().id(),
             context.partitionDirectory(),
-            new ChecksumProvider());
+            new ChecksumProviderRocksDBImpl());
 
     final var submit =
         context.schedulingService().submitActor(snapshotStore, SchedulingHints.ioBound());
